Bug squashed and dependency version updates
• A bug was causing API posts to fail. The problem was that I had declared five columns to the students database table but was only posting four
• Updated Spring Boot  to version 2.7.13 
• Cleaned some old, unnecessary lines from various classes

The next step is to add functionality beyond the current get and post API endpoint capabilities. Put, patch and delete endpoints will be installed. Once the put, patch and delete HTTP methods are implemented I will begin work on the integration tests card to validate the system before moving on to wiring up the front end.