React UI implemented
Next.js for displaying the Spring H2 database implemented. 
Dependencies:
- react
- react-dom
- next
- tailwindcss
- postcss
- autoprefixer
- eslint
- eslint-config-next
- Axios

The decision to use Axios as opposed to the browser default Fetch API was made because Fetch API is not supported in legacy Internet Explorer. This may prevent some end-users from easily completing their transaction. Next for development is the card to add, edit and delete employee data.